refactor: shared layer public api 구조 변경
