+++ client/src/entities/division/api/repository.ts
@@ -1,4 +1,4 @@
-import type { Fetcher } from "@/shared";
+import type { Fetcher } from "@/shared/api";
 import { parseDivisionDto } from "../lib/parse-dto";
 import type { Division, DivisionForm } from "../model/types";
 import type { DivisionCreateDto, DivisionDto, DivisionRepository } from "./types";
