+++ client/src/entities/manual-record/api/repository.ts
@@ -1,4 +1,4 @@
-import type { Fetcher } from "@/shared";
+import type { Fetcher } from "@/shared/api";
 import { parseManualRecordDto } from "../lib/parse-dto";
 import type { ManualRecord, ManualRecordForm } from "../model/types";
 import type { ManualRecordCreateDto, ManualRecordDto, ManualRecordRepository } from "./types";
