+++ client/src/entities/participant/api/repository.ts
@@ -1,4 +1,4 @@
-import type { Fetcher } from "@/shared";
+import type { Fetcher } from "@/shared/api";
 import { parseParticipantDto } from "../lib/parse-dto";
 import type { Participant, ParticipantForm } from "../model/types";
 import type { ParticipantCreateDto, ParticipantDto, ParticipantRepository } from "./types";
