+++ client/src/features/admin-competition/ui/competition-edit-modal.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
+import { Modal, ModalFooter } from "@/shared/ui";
 import type { Competition, CompetitionForm } from "@/entities/competition";
-import { Modal, ModalFooter } from "@/shared";
 
 interface CompetitionEditModalProps {
   isOpen: boolean;
