+++ client/src/features/admin-division/ui/division-create-modal.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
+import { Modal, ModalFooter } from "@/shared/ui"; // shared Public API
 import type { Competition } from "@/entities/competition";
 import type { DivisionForm } from "@/entities/division";
-import { Modal, ModalFooter } from "@/shared"; // shared Public API
 
 interface DivisionCreateModalProps {
   isOpen: boolean;
