+++ client/src/features/admin-division/ui/division-delete-modal.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
+import { Modal, ModalFooter, IconAlert } from "@/shared/ui"; // shared Public API
 import type { Division } from "@/entities/division";
-import { Modal, ModalFooter, IconAlert } from "@/shared"; // shared Public API
 
 interface DivisionDeleteModalProps {
   isOpen: boolean;
