+++ client/src/features/admin-participant/ui/participant-delete-modal.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
+import { Modal, ModalFooter, IconAlert } from "@/shared/ui"; // shared Public API
 import type { Participant } from "@/entities/participant";
-import { Modal, ModalFooter, IconAlert } from "@/shared"; // shared Public API
 
 interface ParticipantDeleteModalProps {
   isOpen: boolean;
