+++ client/src/features/admin-user/ui/user-create-modal.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
+import { Modal, ModalFooter } from "@/shared/ui"; // shared Public API
 import type { UserRegisterForm } from "@/entities/user";
-import { Modal, ModalFooter } from "@/shared"; // shared Public API
 
 type UserFormErrors = Partial<Record<keyof UserRegisterForm, string>>;
