+++ client/src/features/admin-user/ui/user-edit-roles-modal.tsx
@@ -1,7 +1,7 @@
 // src/widgets/edit-user-role-modal.tsx
 import { useState, useEffect } from "react";
+import { Modal, ModalFooter } from "@/shared/ui"; // shared Public API
 import type { User, UserRole } from "@/entities/user";
-import { Modal, ModalFooter } from "@/shared"; // shared Public API
 
 type RoleFormErrors = { roles?: string };
