+++ client/src/features/progress/api/repository.ts
@@ -1,4 +1,4 @@
-import type { Fetcher } from "@/shared";
+import type { Fetcher } from "@/shared/api";
 import { parseProgressDto } from "../lib/parse-dto";
 import type { ProgressState } from "../model/types";
 import type { ProgressDto, ProgressRepository } from "./types";
