+++ client/src/shared/api/index.ts
@@ -1,0 +1,4 @@
+export * from "./errors";
+export * from "./fetcher";
+export * from "./fetcher.authenticated";
+export * from "./fetcher.fetch";
