--- client/src/shared/index.ts
@@ -1,12 +1,0 @@
-export * from "./api/errors";
-export * from "./api/fetcher";
-export { FetchApiFetcher } from "./api/fetcher.fetch";
-export { AuthenticatedFetcher, type SessionProvider } from "./api/fetcher.authenticated";
-export * from "./ui/base-modal";
-export * from "./ui/modal";
-export * from "./ui/modal-footer";
-export * from "./ui/icon-x";
-export * from "./ui/icon-alert";
-export * from "./lib/use-real-time-timer";
-export * from "./lib/date-utils";
-export { cn } from "./lib/utils";
