+++ client/src/shared/ui/index.ts
@@ -1,0 +1,5 @@
+export * from "./base-modal";
+export * from "./icon-alert";
+export * from "./icon-x";
+export * from "./modal";
+export * from "./modal-footer";
